style(TeamPage): стилизация страницы списка актеров
