+++ frontend/src/components/ui/Shared/MainTitle/main-title.scss
@@ -1,5 +1,3 @@
 .main-title{
-    font-weight: 400;
-    font-size: 36px;
-    text-align: center;
+
 }
