+++ frontend/src/components/ui/TeamPage/TeamCardBrief/team-card-brief.scss
@@ -1,6 +1,23 @@
 .team-card-brief{
-    width: 295px;
-    height: 422px;
-    background-color: var(--skeleton);
-    // placeholder
+    display: flex;
+    max-width: 250px;
+    flex-direction: column;
+    gap: 12px;
+
+    &__photo{
+        max-width: 250px;
+        object-fit: cover;
+        height: 350px;
+    }
+    &__name{
+        font-size: 24px;
+        line-height: 28px; 
+    }
+    &__position{
+        margin-top: 2px;
+        color: var(--gray-basic);
+
+        font-size: 18px;
+        line-height: 22px; 
+    }
 }
