+++ frontend/src/components/ui/TeamPage/TeamFilter/team-filter.scss
@@ -1,7 +1,4 @@
 .team-filter{
-    margin-top: 40px;
-
     display: flex;
     justify-content: center;
-    gap: 30px;
 }
