+++ frontend/src/components/ui/TeamPage/TeamList/team-list.scss
@@ -1,10 +1,8 @@
 .team-list{
-    margin-top: 60px;
-    margin-bottom: 20px;
- 
-
     display: flex;
+ 
+    gap: 40px;
+    align-self: stretch;
     flex-wrap: wrap;
-    justify-content: center;
-    gap: 40px
+    justify-content: space-around;
 }
