+++ frontend/src/components/ui/TeamPage/team-page.scss
@@ -1,6 +1,12 @@
 .team-page{
-
-    &__title{
-        margin-top: 120px;
+    margin-bottom: 100px;
+    padding: 0 50px;
+    &__header{
+        margin-top: 100px;
+        margin-bottom: 50px;
+        display: flex;
+        justify-content: space-between;
+        align-items: center
+        
     }
 }
