+++ frontend/src/pages/TeamPage/TeamPage.tsx
@@ -1,6 +1,19 @@
-import { Team } from '@components/TeamPage';
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
+import { setCategory } from 'services/slices/teamSlice';
+
+import { TeamPageUI } from '@components/ui/TeamPage';
+import { AppDispatch } from '@services/store';
+import { TeamFilterCategories } from 'consts';
 
 export const TeamPage: React.FC = () => {
-    return <Team />;
+    const dispatch = useDispatch<AppDispatch>();
+    const location = useLocation();
+
+    useEffect(() => {
+        dispatch(setCategory(TeamFilterCategories.Actors));
+    }, [location, dispatch]);
+
+    return <TeamPageUI />;
 };
